+++ src/Merzbild.jl
@@ -55,7 +55,7 @@ export merge_roulette!
 export NNLSMerge, compute_multi_index_moments, merge_nnls_based!
 export merge_nnls_based_rate_preserving!
 export ElectronNeutralInteractions, ComputedCrossSections
-export load_electron_neutral_interactions, create_computed_crosssections, DataMissingException
+export create_computed_crosssections, DataMissingException
 export ScatteringIsotropic, ScatteringOkhrimovskyy
 export ElectronEnergySplitEqual, ElectronEnergySplitZeroE
 export CSExtendZero, CSExtendConstant