+++ src/collisions/collision_cross_sections.jl
@@ -418,10 +418,10 @@ end
 
 
 """
-    load_electron_neutral_interactions(species_data, filename, databases, scattering_laws, energy_splits)
+    ElectronNeutralInteractions(species_data, filename, databases, scattering_laws, energy_splits)
 
 Load electron-neutral interaction data from an LXCAT format XML file for
-a set of given neutral species.
+a set of given neutral species and instantiate a `ElectronNeutralInteractions` instance.
 
 # Positional arguments
 * `species_data`: vector of `Species` data for the neutral species
@@ -437,7 +437,7 @@ of the cross-section database in the XML file to use for the species
 # Throws
 `DataMissingException` if data not found or not all required data present.
 """
-function load_electron_neutral_interactions(species_data, filename, databases, scattering_laws, energy_splits)
+function ElectronNeutralInteractions(species_data, filename, databases, scattering_laws, energy_splits)
     neutral_indexer::Vector{Int64} = []
     mass_ratios::Vector{Float64} = []
     elastic_cs_vector::Vector{ElasticScattering} = []